requirements.txt
Django==3.2.6
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
psycopg2-binary==2.9.1
xmltodict==0.12.0
